app/requirements.txt
telethon==1.36.0
httpx==0.27.0
PyYAML==6.0.2
